Split top-level help rendering into its own lightweight module and trim help.ts down to command-help fallback only, reducing residual --help and --version startup cost
Result: {"status":"keep","total_ms":650.145}
