+++ autoresearch.jsonl
@@ -29,3 +29,4 @@
 {"run":25,"commit":"5376ee4","metric":1266.005,"metrics":{},"status":"keep","description":"Extend the static direct-render help path to auth, login, logout, and status so the full help workload avoids heavy command loading across the auth family too","timestamp":1776388371779,"segment":4,"confidence":null,"iterationTokens":4311,"asi":{"hypothesis":"The auth-family help commands still dominate the widened benchmark because they go through the heavy commands/help path; static direct-render help should collapse them the same way as other admin commands.","secondary_metrics":{"help_ms":97.765,"agent_help_ms":141.095,"install_help_ms":93.647,"skills_help_ms":108.298,"sync_help_ms":99.655,"mcp_help_ms":100.403,"token_help_ms":83.528,"cache_help_ms":80.486,"auth_help_ms":92.032,"login_help_ms":86.829,"logout_help_ms":96.184,"status_help_ms":88.958,"version_ms":97.124,"stdout_bytes":49035,"passes":1},"learned":"Static help now generalizes across the whole admin/auth family: the full benchmark improved by about 67% vs the widened-auth baseline, and the formerly worst auth/help paths dropped from ~588-701ms down to ~87-96ms."}}
 {"run":26,"commit":"18805ad","metric":850.19,"metrics":{},"status":"keep","description":"Move all benchmarked command help metadata into the lightweight static-help module so agent, admin, and auth help paths all render directly without importing heavy command/help registries","timestamp":1776388544984,"segment":4,"confidence":7.217745872563518,"iterationTokens":10585,"asi":{"hypothesis":"If the benchmarked agent help path also uses the same direct-render static help module as admin/auth commands, the full help workload should collapse further without the chunk-graph regression seen in the separate agent-spec renderer approach.","secondary_metrics":{"help_ms":68.558,"agent_help_ms":78.481,"install_help_ms":64.247,"skills_help_ms":53.45,"sync_help_ms":47.656,"mcp_help_ms":55.932,"token_help_ms":90.052,"cache_help_ms":54.615,"auth_help_ms":59.437,"login_help_ms":61.186,"logout_help_ms":73.859,"status_help_ms":73.925,"version_ms":68.793,"stdout_bytes":49035,"passes":1},"learned":"Unifying benchmarked help metadata behind one direct-render static module works much better than separate agent/admin renderers: the full workload fell another ~33% vs the widened-auth baseline, and all command help paths now cluster roughly in the 48-90ms range."}}
 {"type":"config","name":"Speed up octocode-cli help startup (all agent/admin/auth help)","metricName":"total_ms","metricUnit":"ms","bestDirection":"lower"}
+{"run":27,"commit":"ebd56a0","metric":844.839,"metrics":{},"status":"keep","description":"Baseline for full help workload covering top-level help, all benchmarked agent helps, admin/auth helps, and version","timestamp":1776388642879,"segment":5,"confidence":null,"iterationTokens":106,"asi":{"hypothesis":"Widen the workload to all benchmarked command help flows so remaining optimizations are measured against broad real coverage rather than one representative agent command.","secondary_metrics":{"help_ms":43.473,"search_code_help_ms":45.429,"get_file_help_ms":44.792,"view_structure_help_ms":48.936,"search_repos_help_ms":45.365,"search_prs_help_ms":42.848,"package_search_help_ms":43.627,"install_help_ms":42.587,"skills_help_ms":45.674,"sync_help_ms":52.645,"mcp_help_ms":51.396,"token_help_ms":47.735,"cache_help_ms":52.611,"auth_help_ms":54.117,"login_help_ms":49.681,"logout_help_ms":41.548,"status_help_ms":49.214,"version_ms":43.16,"stdout_bytes":76095,"passes":1},"learned":"After broad static-help coverage, all benchmarked command help paths are already tightly clustered around ~42-54ms. Remaining opportunity is mostly top-level help/version, not per-command help."}}
