+++ packages/octocode-cli/src/cli/help.ts
@@ -1,101 +1,6 @@
 import { c, bold, dim } from '../utils/colors.js';
 import type { CLICommand } from './types.js';
 
-declare const __APP_VERSION__: string;
-
-export function showHelp(): void {
-  const lines = [
-    '',
-    `  ${c('magenta', bold('🔍🐙 Octocode CLI'))} - Install and configure octocode-mcp`,
-    '',
-    `  ${bold('USAGE')}`,
-    `    ${c('magenta', 'octocode-cli')} [command] [options]`,
-    '',
-    `  ${bold('AGENT TOOLS')}`,
-    `    ${c('magenta', 'search-code')}      Search code in GitHub repositories`,
-    `    ${c('magenta', 'get-file')}         Fetch file content from a GitHub repo`,
-    `    ${c('magenta', 'view-structure')}   View directory structure of a GitHub repo`,
-    `    ${c('magenta', 'search-repos')}     Search GitHub repositories`,
-    `    ${c('magenta', 'search-prs')}       Search GitHub pull requests`,
-    `    ${c('magenta', 'package-search')}   Search npm or Python packages`,
-    '',
-    `  ${bold('SETUP & ADMIN')}`,
-    `    ${c('magenta', 'install')}          Configure octocode-mcp (alias: setup)`,
-    `    ${c('magenta', 'skills')}           Install Octocode skills across AI clients`,
-    `    ${c('magenta', 'cache')}            Inspect and clean Octocode cache directories`,
-    `    ${c('magenta', 'sync')}             Sync MCP configurations across all IDEs`,
-    `    ${c('magenta', 'mcp')}              Manage MCP marketplace non-interactively`,
-    `    ${c('magenta', 'auth')}             Manage GitHub authentication`,
-    `    ${c('magenta', 'login')}            Authenticate with GitHub`,
-    `    ${c('magenta', 'logout')}           Sign out from GitHub`,
-    `    ${c('magenta', 'status')}           Show GitHub authentication status`,
-    `    ${c('magenta', 'token')}            Print the stored GitHub OAuth token`,
-    '',
-    `  ${bold('OPTIONS')}`,
-    `    ${c('cyan', '-h, --help')}       Show this help message`,
-    `    ${c('cyan', '--tools-context')}  Print Octocode MCP instructions and tool schemas`,
-    `    ${c('cyan', '--tool <name> <json>')}  Run one Octocode tool with one JSON payload`,
-    `    ${c('cyan', '-v, --version')}    Show version number`,
-    '',
-    `  ${bold('EXAMPLES')}`,
-    `    ${dim('# Search code in a GitHub repo')}`,
-    `    ${c('yellow', "octocode-cli search-code --query 'useReducer dispatch' --owner facebook --repo react")}`,
-    '',
-    `    ${dim('# Fetch a file around a matched string')}`,
-    `    ${c('yellow', 'octocode-cli get-file --owner facebook --repo react --path packages/react/src/React.js --match-string useState')}`,
-    '',
-    `    ${dim('# Bulk search via stdin')}`,
-    `    ${c('yellow', `echo '{"queries":[...]}' | octocode-cli search-code`)}`,
-    '',
-    `    ${dim('# Interactive installer')}`,
-    `    ${c('yellow', 'octocode-cli install')}    ${dim('(alias: setup)')}`,
-    '',
-    `    ${dim('# Non-interactive install for Cursor')}`,
-    `    ${c('yellow', 'octocode-cli install --ide cursor --method npx')}`,
-    '',
-    `    ${dim('# Install for Claude Desktop using direct method')}`,
-    `    ${c('yellow', 'octocode-cli install --ide claude-desktop --method direct')}`,
-    '',
-    `    ${dim('# Check GitHub authentication')}`,
-    `    ${c('yellow', 'octocode-cli auth')}`,
-    '',
-    `    ${dim('# Get token from Octocode (default)')}`,
-    `    ${c('yellow', 'octocode-cli token')}`,
-    '',
-    `    ${dim('# Get token from gh CLI')}`,
-    `    ${c('yellow', 'octocode-cli token --type=gh')}`,
-    '',
-    `    ${dim('# Sync MCP configs across all IDEs')}`,
-    `    ${c('yellow', 'octocode-cli sync')}`,
-    `    ${c('yellow', 'octocode-cli sync --status')}`,
-    '',
-    `    ${dim('# MCP marketplace via CLI')}`,
-    `    ${c('yellow', 'octocode-cli mcp list --search browser')}`,
-    `    ${c('yellow', 'octocode-cli mcp install --id playwright-mcp --client cursor --force')}`,
-    `    ${c('yellow', 'octocode-cli mcp remove --id playwright-mcp --client cursor')}`,
-    '',
-    `    ${dim('# Install Octocode skills')}`,
-    `    ${c('yellow', 'octocode-cli skills install')}`,
-    `    ${c('yellow', 'octocode-cli skills install --skill octocode-researcher')}`,
-    `    ${c('yellow', 'octocode-cli skills install --targets claude-code,cursor,codex --mode symlink')}`,
-    '',
-    `    ${dim('# Remove one installed skill')}`,
-    `    ${c('yellow', 'octocode-cli skills remove --skill octocode-researcher --targets claude-code,cursor')}`,
-    '',
-    `    ${dim('# Run a tool directly')}`,
-    `    ${c('yellow', `octocode-cli --tool localSearchCode '{"path":".","pattern":"runCLI"}'`)}`,
-    `    ${c('yellow', `octocode-cli --tool githubSearchCode '{"owner":"bgauryy","repo":"octocode-mcp","keywordsToSearch":["tool"]}'`)}`,
-    '',
-    `    ${dim('# Print MCP instructions and tool schemas')}`,
-    `    ${c('yellow', 'octocode-cli --tools-context')}`,
-    '',
-    c('magenta', `  ─── 🔍🐙 ${bold('https://octocode.ai')} ───`),
-    '',
-  ];
-
-  process.stdout.write(`${lines.join('\n')}\n`);
-}
-
 export function showCommandHelp(command: CLICommand): void {
   const lines = [
     '',
@@ -126,10 +31,4 @@ export function showCommandHelp(command: CLICommand): void {
   }
 
   process.stdout.write(`${lines.join('\n')}\n`);
-}
-
-export function showVersion(): void {
-  const version =
-    typeof __APP_VERSION__ !== 'undefined' ? __APP_VERSION__ : 'unknown';
-  console.log(`octocode-cli v${version}`);
 }
